SignView & Modal UI + logic improvements
Add support for gestures in modal view. The gestures are sent from the SignView into the Modal component (thru megablob action).

Signview knows when the modal is open and sends the interpreted gestures to that view. In addition SignView now displays the Transcripts that are spoken (by Alexa) after the sign-sentance is sent to the chat room.

Note there is only one chat-room available, Alexa, that cannot be changed in the current implementation.